+++ dsptoolbox/classes/_filter.py
@@ -288,7 +288,7 @@ def _filter_on_signal(
                 "Filter output is complex. Imaginary part is saved in "
                 + "Signal as time_data_imaginary"
             )
-        new_time_data = new_time_data.astype("cfloat")
+        new_time_data = new_time_data.astype(np.complex128)
 
     # Create new signal
     new_time_data[:, channels] = y
@@ -395,7 +395,7 @@ def _filter_on_signal_ba(
                 "Filter output is complex. Imaginary part is saved in "
                 + "Signal as time_data_imaginary"
             )
-        new_time_data = new_time_data.astype("cfloat")
+        new_time_data = new_time_data.astype(np.complex128)
 
     # Create new signal
     new_time_data[:, channels] = y
